# api-gateway/docs/swagger.yaml
openapi: 3.0.3
info:
  title: Auction System API
  version: 1.0.0
  description: Production-grade auction platform — all endpoints
servers:
  - url: http://localhost:8080
    description: Local (API Gateway)
...
paths:
...
  /api/v1/search:
    get: {tags: [Search], summary: Search auctions, parameters: [{name: q, in: query, schema: {type: string}}, {name: category, in: query, schema: {type: string}}, {name: min_price, in: query, schema: {type: number}}, {name: max_price, in: query, schema: {type: number}}, {name: sort_by, in: query, schema: {type: string, enum: [price_asc,price_desc,newest,ending_soon,most_bids]}}], responses: {'200': {description: Results}}}
  /api/v1/search/suggest:
    get: {tags: [Search], summary: Autocomplete, parameters: [{name: q, in: query, required: true, schema: {type: string}}], responses: {'200': {description: Suggestions}}}
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT